Added: watch_markdown
A script that periodcally (every 0.5 seconds) checks if a file was modified.
When it was then it calls `pandoc` to create a PDF file for it. It also
automatically starts evince to watch the PDF.

Can for example be used to open a file with `gedit` and whenever the file
is saved it will re-generate the PDF causing `evince` to update the file.