First version of finding duplicates
Group files by size and then computes SHA512 over first 4 KiB of file
content, prints the results of what is left. Results can be duplicates
but are not safe as only first 4 KiB is compared.

But as it only shows results so far, its okay. Further addition is maybe
de-duplication with symlinks, file-glob pattern which files to check
and file-size restriction (minimum, maximum).