Use read_bytes from Sq
This now seems more complicated and sometimes is a feature. Opening
a file can fail. That's how it is. So calling read_bytes()
return an Result. At this point i need to decide what should happen.

Here i now decide to throw an exception. Instead that an exception is
thrown for me. Also could have printed a warning or do something else.